rtc: stm32_rtc: fix fraction difference computation
stm32_rtc_get_second_fraction() returns an unsigned value.
Cast both fractions values (current and reference) before computing
the difference, in order to take negative values into account.

Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/151364

+++ core/drivers/stm32_rtc.c
@@ -233,11 +233,11 @@ static uint32_t stm32_rtc_get_second_fraction(struct stm32_rtc_calendar *cal)
  * This function computes the fraction difference between two timestamps.
  * Here again the returned value is in milliseconds.
  ******************************************************************************/
-static unsigned long long stm32_rtc_diff_frac(struct stm32_rtc_calendar *cur,
-					      struct stm32_rtc_calendar *ref)
+static signed long long stm32_rtc_diff_frac(struct stm32_rtc_calendar *cur,
+					    struct stm32_rtc_calendar *ref)
 {
-	return stm32_rtc_get_second_fraction(cur) -
-		stm32_rtc_get_second_fraction(ref);
+	return (signed long long)stm32_rtc_get_second_fraction(cur) -
+	       (signed long long)stm32_rtc_get_second_fraction(ref);
 }
 
 /*******************************************************************************
